Random array generator version 1.
This program will automate the testing process and allow us to be able to check the behavior of the program for large test cases. Version 1 will allow the tester to choose whether the array in in ascending or random order. The next version will replace the array with a linked list as we will be generating negative numbers, thus the size of the test cases will need to be changed dynamically. As well as making sure that there is no repetition in the array.